rgw/beast: (non-functional) refactor Connection and ConnectionList to header
Just like rgw_asio_frontend_timer.h, factor out the Connection struct and
ConnectionList class from the anonymous namespace in rgw_asio_frontend.cc
to a new header rgw_asio_frontend_connection.h (under namespace rgw::asio).

This is especially needed to get better coverage thru unittesting.

This is non-functional: no logic changes, the types, methods, and behavior
are identical as before. The frontend .cc uses type aliases to maintain the
same unqualified names.

(cherry picked from commit 222a39366d27aaadf2c96a33b8a29275f233afab)
Fixes: https://tracker.ceph.com/issues/78306
